backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Neo4j graph database driver
neo4j==5.26.0

# Data ingestion & parsing
pandas==2.2.3
openpyxl==3.1.5

# Data validation
pydantic[email]==2.10.3

# Authentication (JWT)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File upload support
python-multipart==0.0.12

# ML scoring
scikit-learn==1.6.0
joblib==1.4.2

# Graph analysis
networkx==3.4.2

# Environment variable loading
python-dotenv==1.0.1
